fixed: re-added file operation checks that were gone after files_in_lib branch merge

+++ xbmc/video/windows/GUIWindowVideoBase.cpp
@@ -1479,7 +1479,9 @@ void CGUIWindowVideoBase::OnDeleteItem(CFileItemPtr item)
       return;
   }
 
-  CFileUtils::DeleteItem(item);
+  if (g_guiSettings.GetBool("filelists.allowfiledeletion") &&
+      CUtil::SupportsFileOperations(item->m_strPath))
+    CFileUtils::DeleteItem(item);
 }
 
 void CGUIWindowVideoBase::MarkWatched(const CFileItemPtr &item, bool bMark)
